pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asmreader"
version = "0.1.0"
description = "Turn any text into a soothing ASMR-style reading experience"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Autoscatto", email = "4utoscatto@gmail.com" },
]
keywords = ["asmr", "text-to-speech", "tts", "pdf", "reader"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing :: General",
    "Typing :: Typed",
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "ebooklib>=0.18",
    "kokoro-onnx>=0.3.5",
    "lxml[html-clean]>=5.3.0",
    "markdown>=3.7",
    "pypdf>=3.17.1",
    "python-magic>=0.4.27",
    "readability-lxml>=0.8.1",
    "requests>=2.32.3",
    "simpleaudio>=1.0.4",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.0",
    "tdqm>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.24.1",
    "ruff>=0.2.1",
    "black>=24.1.1",
    "mypy>=1.8.0",
    "types-tqdm>=4.67.0.20241221",
    "types-Markdown",
    "types-requests",
    "types-beautifulsoup4",
]

[project.scripts]
asmreader = "asmreader.cli:main"
download_model = "asmreader.download_model:main"  # Entry point for the download_model script

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "ANN", "S", "A", "C4", "T20", "RET", "SIM", "PTH"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # Allow assert in tests

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=asmreader"

[tool.hatch.build.targets.wheel]
packages = ["asmreader"]

[tool.hatch.build.targets.sdist]
include = [
    "/asmreader",
    "/tests",
    "README.md",
    "LICENSE"
]
